Unibook app
As part of the mobile application term project, we tried to make the unibook platform for people who want to sell second hand books or people who want to buy second hand books. The application is based on android and firebase and now Unibook is working very simply and there are many deficiencies already, but it is open to development.